refactor: particle distance float

+++ include/config.cuh
@@ -57,7 +57,7 @@ struct Configuration {
 //    FixedHSLA background{ IC_16b, icenc(0.06), icenc(0.15), ICE_1 };
     const char * output = "plot.mp4";
     const char * particles_file = nullptr;
-    unsigned long particle_distance = 10;
+    float particle_distance = 10;
     unsigned long margin = 4;
     unsigned long lloyd_iterations = 8;
     ExecutionMode mode = ExecutionMode::GPU;
